backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
google-generativeai>=0.8.0
supabase>=2.9.0
python-dotenv>=1.0.1
httpx>=0.27.0
aiofiles>=23.2.1
pydantic>=2.10.0
python-jose[cryptography]>=3.3.0
